Add script to sync code from upstream TF repo.
Currently, this script copies the shared TFL/TFLM code as well as the
entire micro sub-folder in Tensorflow. Once TFLM development switches to
using the current repo, the sync script will still be used to get the
shared TFL/TFLM code.

Also, updated the BUILD file to use glob so that there is one less place
to change when a new kernel is ported from TFL to TFLM.

Tested:
 * ran the sync script (which updated reduce.h):
     ```
     ci/sync_from_upstream_tf.sh
     ```
 * confirmed that the docker container can run all the tests
   successfully:
     ```
     docker build -t tflm-test -f ci/Dockerfile.micro .
     docker run --rm tflm-test
     ```