Also add any new files to the sync PR.
Without the additional step of `git add`, any untracked files will not
be part of the PR that is created with the sync workflow.

Manually tested this change on my fork of tensorflow by removing a file,
pushing it to main and then confirming that the sync workflow does not
add to the PR prior to this change, but does add it to the PR after this
change.