patch TFLM visibility updates from #73 prior to sync from upstream TF.
These visibility changes are specific to the TFLM code in
github.com/tensorflow/tflite-micro and would be awkward to upstream to
github.com/tensorflow/tensorflow.

Additionally, we are getting close to switching the current repository
to be the source of truth for TFLM code and have thus made the decision
to temporarily add some patches during the sync from upstream TF to the
current repo.

Manually tested that running `ci/sync_from_upstream_tf.sh` with the
current change avoids reverting the changes from #73 (such as with #76).