Have the project generation script explicitly call make third_party_downloads.
This is required only because some of the third_party_downloads are
still using the Makefile functions and thus require an explicit
`make third_party_downloads`.

Additionally, we sometimes also want to avoid all downloads and this PR
adds in a flag for that as well.

Fixes #119 and the corresponding internal bug http://b/184886633

Related bugs are http://b/143904317 and http://b/183229719