Explicitly use python3 from the Makefile.
For the Xtensa docker container, we were getting an error message along
the lines of `python not found`. And the result was that the
specialization was happening incorrectly (i.e. the specialization for
the kernels was failing since it relies on python since #160).

Exactly why this error passed the CI for #160 but started failing after
is not entrely clear but is something that we are not going to
investigate.