Remove `build_test` from the BUILD files.
Since the CI builds all the targets (binary and test), and the source of
truth for all the TFLM code is having a build_test is not adding any additional value.

Additionally, a build_test seems to add additional downloads which would
make it slower to run the bazel CI from within a docker container.