Add a new test case for conv operator (#329)
This new test case is based on issue #329 to increase
coverage on optimized kernel's data precision for conv operator.
In this test, input, output and filter are all 8 bits and filter
tensor is of dimension 8x3x3x3 with different scales per output channel.

TESTED= local test with x86 and HiFi4.