Have make fail if python script has an error.
Prior to this change, the generate_cc_arrays.py script could fail (for
example because PIL not being installed) but the error would be
silently ignored. Instead, the build would fail much later because of
missing header files and the root cause would be hard to debug.

Follow-up PR will also use this error handling approach for other places
where we call into Python and bash scripts from the make.

BUG=http://b/196630709