Remove -D<target>
There isn't any need to have a special define for the target. In fact,
having such a define can create trouble when the TARGET and
OPTIMIZED_KERNEL_DIR have the same name.

Priot to this change, -D<target> and -D<optimized_kernel_dir> could
result in the same define which would interfere with the kernel variant
definitions (which are tied to the optimized_kernel_dir and are
independent of the target).

Manually tested that the following command fails without this change, and passes with it:
```
make -f tensorflow/lite/micro/tools/make/Makefile TARGET=xtensa TARGET_ARCH=hifi4 XTENSA_CORE=F1_190305_swupgrade -j8 keyword_benchmark
```

BUG=http://b/197889069