Call test_size.sh to produce size comparison report in test_xtensafusion_f1.sh and change test_size.sh to take more makefile arguments.
The reason to piggy back to test_xtensa_fusion_f1.sh script
is to avoid loading the docker again in presubmit.

This change will lead to produce a size comparison report
at fusion_f1 presubmit test, but will not lead to failure
in presubmit yet if the size increase. Presubmit failure
on size increase will be only be enabled later after more tests.

keyword_benchmark is chosen as the size bench mark target.

BUG=http://b/196637015