Refactor MicroAllocator such that a MemoryPlanner can be injected into the MicroAllocator
This is the first step to allow offline memory planner.

When a MemoryPlanner is not injected at creation time for
a MicroAllocator, a default GreedyMemoryPlanner is created
from the arena. This change is hence backward compatible
at the MicroAllocator api level. The arena usage slightly increase
by 80 bytes on linux x86 due to the GreedyMemoryPlanner is moved from
stack to the arena in such case.

This corresponds to internal cl/395749105

BUG=http://b/199230276