Add flatbuffer_utils and corresponding test to the bazel build.
This is the first step towards being able to manipulate flatbuffers from
Python from the TFLM github repository.

Manually confirmed that the following command works:
```
bazel test tensorflow/lite/tools:flatbuffer_utils_test
```

Note that the code that is sync'd from TF needs to have an updated
import path and we have changed the sync script to perform this
transformation.

These python tests will also be run as part of the bazel CI.

BUG=http://b/204109200