Add int32_t and int8_t support to the FILL op and avoid unconditional check on the dims tensor.
Some tflite model requires int32 and int8 FILL OP and additionally, at
init and prepare time,
values of the dims tensor may not be specified yet and shall not be used
for validity
check unconditionally.

Also refactor fill_test.cc to have a factory method to create
KernelRunner for the FILL op test.

This correspond to internal cl/406052107

BUG=http://b/204237320