Fix erroneous write from EXPAND_DIMS to an array that can be in read-only region
TFLM actually point to the tflite model buffer for the dimension array
of a tensor. However, EXPAND_DIMS also try to
write to this dimension array, which leads to segfault when the model
buffer is
declared as constant and in RO data section.
One such github issue is
https://github.com/tensorflow/tflite-micro/issues/597.

The fix is to avoid writing altogether since TFLM fundamentally assumes
fixed dimension for tensors.

BUG=https://b/204237320