Use csv file name to determine type of the array.
This will allow us to have either int8 or int16 be generated from csv
files and will expand the types of integration tests that we can run.

Renamed all the files with:
```bash
for CSV in $(ls *.csv)
do
  mv "${CSV}" "$(basename ${CSV} .csv)_int16.csv"
done
```

Also,
  * added the integration_test to x86 makefile per-PR CI
  * changed bazel target to cc_test so that the integration test is run
    as part of CI.

If this slows down the CI, we will add a mechanism to run a more
exhaustive test suite nightly, but we are currently going to err on the
side of having more coverage.

For now, running the integration tests on x86 is fast.

BUG=http://b/205042963