Remove all hardware-specific code from the TFLM examples.
In addition, this PR also removes support for the Makefile based project generation.

Relevant issues:
 * https://github.com/tensorflow/tflite-micro/issues/407
 * https://github.com/tensorflow/tflite-micro/issues/445

The documentation, tests for the examples etc. still need to be updated,
and that will be addressed in follow-up PRs.

Additionally, there are some existing users that may not have fully
transitioned over to the new project generation and examples guidance.
This will be a breaking change for these users but is something that we
need to do to make progress on the goals of #407.

BUG=#407