Replace magic constants in micro_allocator_test.cc with readable expressions.
This change replaces the magic constants in micro_allocator_test.cc
with a more readable expression.
This is a baby first step towards a better code base and example
of estimating the arena usage.

Additionally, two functions in micro_allocator_test.cc is replaced
with better names that reflect their functionality.

BUG=http://b/207157610