Replace part of magic number with expression in memory_arena_threshold_test.cc
Similar to MicroAllocator, define a constant for the default tail
usage for RecordingMicroAllocator and use it in
memory_arena_threshold_test.cc.
Some original expression only account for memory planner size and this
change account for micro allocator as well.
This is another baby step for cleaning up magic number that make test
brittle.

BUG=http://b/207157610