Add Unused() for MicroPrintf to avoid #if on TF_LITE_STRIP_ERROR_STRINGS on different files
This is a different fix than pr #718 that avoids putting
In current form, if we add any code like below, it will lead to a unused variable
error unless we wrap it around #if.

```
int variable_for_print = function_for_print();
MicroPrintf("This leads to compiler error %d", variable_for_print);
```

Verify that this change does not increase size by adding a bunch of
MicroPrintf in hello_world example and observe the size of binary
remain unchanged.

BUG=http://b/208249141