Add base helpers for using Pigweed's unit test module
Add an include file and Makefile function to be used when
declaring unit tests using Pigweed's unit test module, mirroring
the way unit tests are declared for micro_test. The Makefile
function adds the include directories and objects in Pigweed on
which unit tests depend.

Add a common main() to be linked with all Pigweed-based unit
tests. main() runs all tests defined in the program. (The tests
are located automatically via registrations made during static
initialization).

The common include file adds an assertion macro EXPECT_NEAR,
which Pigweed does not have natively, but is found in micro_test
and used in several places in tflite-micro. There may yet be other
incompatibilities to be adapted in a similar manner.