This reverts #839 and #846.
Applyling labels via github workflow on PR approval ended up being
fairly complex to implement and suffers from a race condition as
decribed in #839.

Instead, we are going to implement the same functionality using a
centralized TensorFlow bot account and GitHub App which is much more
suited to what we need.

BUG=issuetracker.google.com/issues/210946861