Add a DeallocateTemp to pair with AllocateTemp in SimpleMemoryAllocator. (#868)
* Add a DeallocateTemp to pair with AllocateTemp in SimpleMemoryAllocator.

Add a DeallocateTemp to pair with AllocateTemp in SimpleMemoryAllocator.

This is a backport of internal cl/423466958

The idea is to pair DeallocateTemp with existing AllocateTemp
to provide a better memory auditing for temporary buffer.

After all existing AllocateTemp have been paired with a DeallocateTemp,
we can then enable an error check when the entire temp memory region
is to be reclaimed, but there are still outstanding temp buffers.

Currently, GetInput api use AllocateTemp under the hood and requires a
separate CL to change the api before we can enforce the assertion.

See go/tflm-pairing-allocatetemp-with-deallocatetemp  for more info

BUG=https://b/209453859

* Using shorter ^= expression

Co-authored-by: Ting Yan <94130036+tingyan19@users.noreply.github.com>