Add AllocateTempInputTensor to replace GetInput() and Add AllocateTempOutputTensor to replace GetOutput()
This corresonds to internal cl/424648139

Add more informative api AllocateTempInputTensor to replace GetInput()
and AllocateTempOutputTensor to replace GetOutput(). These two new API
call out the fact that what GetInput and GetOutput returns are in
temporary buffer and  need to call DeallocateTempTensor() afterwards.

IF operator is the first OP to use these APIs.

BUG=https://b/209453859