Clean up GetInput/GetOutput api usage in squeeze, transpose, strided_slice op
Clean up api usage in squeeze, transpose, strided_slice op by replacing
GetInput/GetOutput with AllocateTempXXX/DeallocateTempXXX in micro
context api in squeeze, transpose, strided_slice op

Additionally, squeeze and transpose is using TfLiteTensor in Invoke,  which
is againts the recommendation and unnecessary. Replace them with the usage of
TfLiteEvalTensor.

This is a continuation of pr #894 to improve both readability and safety of getting tensors.

The next step is to clean the GetInput/GetOutput in test_helpers.cc
before enforcing the pair of AllocateTempXXX and DeallocateTemp globally.

BUG=https://b/209453859