Rename the generic SetHeadBufferSize() to two separate functions
This mirrors internal cl/432564758.

Rename the generic SetHeadBufferSize() to two separate functions
ResizeBuffer() and ReserveNonPersistentOverlayMemory().
This is to highlight the real purpose of SetHeadBufferSize() in
different places and remove the reference to Head to abstract out the implementation detail.

This is the first step to abstract out the interface function from
SimpleMemoryAllocator so that we can eventually support separate NonPersistent and Persistent
arena, in which case,the Head is not meaningful.

BUG=https://b/219833563