Add int8 and float SVDF for non-HiFi Xtensa (#999)
* Add int8 and float SVDF for non-HiFi Xtensa

The Xtensa accelerated SVDF implementation is only supported on HiFi, so
it falls back to the reference implementation on VP6. This was hardcoded
to an int16 implementation in xtensa/svdf.cc, as the HiFi accelerated
implementation only supports int16. This PR enables float and int8
support when falling back on the reference implementation. This is only
impacts non-HiFi Xtensa cores, such as the VP6

BUG=b/225235693

* Replace TF_LITE_KERNEL_LOG with MicroPrintf

Co-authored-by: Ting Yan <94130036+tingyan19@users.noreply.github.com>