Add conv kernel unit tests
* Add conv kernel test with 3x3 filter, 2x2 stride, and 4x4 EVEN size
  input (padding 'same').
* Add conv kernel test with 3x3 filter, 2x2 stride, and 5x5 ODD size
  input (padding 'same').

Both tests pass for the TFLM reference kernel for conv when run with:

$ make -f tensorflow/lite/micro/tools/make/Makefile \
  test_kernel_conv_test

When run for the xtensa kernel with:

$ make -f tensorflow/lite/micro/tools/make/Makefile TARGET=xtensa \
  TARGET_ARCH=vision_p6 OPTIMIZED_KERNEL_DIR=xtensa \
  XTENSA_CORE=${VP6_CORE} test_kernel_conv_test

The 4x4 even-size input fails and the 5x5 odd size input passes.

Note that the 4x4 test is currently excluded from the Xtensa tests by an
include guard. It must be included to reproduce the failure.

BUG=b/228102789