Add GetOptionalTensorData API
BUG=b/232005813

Bias tensors can be optional for certain ops. The existing GetTensorData
API is modified to assert upon receiving a nullptr, optional tensors
should be retrieved via GetOptionalTensorData now. The new API should
make it explicit that a tensor is optional and so check its return
value. Only reference kernels are updated in this PR.