Mark all subgraph inputs as used in the memory allocator
This fixes a bug when there are multiple subgraphs and one of the non-main subgraphs (e.g. through a while loop) has inputs that aren't used. The memory planner did not mark these inputs as used, but the while-loop implementation does make copies from one subgraph to another. Without this fix it would cause writes to other memory addresses, resulting in incorrect results if unlucky.