ifdef out error reporter member when building without error strings
When building without error strings, there there are no references to error reporter, which makes the Xtensa clang compiler
(and perhaps others) throw an "Unused Private Field" error. Although the
makfile defines -Wno-unused-private-field in release build, when building position independent code, xt-clang, for no apparent reason, insists on throwing an error. This seemse to be the only way to avoid the error.

BUG=228871380