Add armclang support for target based on Arm Corstone-300 software
Armclang need to use -ffp-mode=full because quantization unit test is
using NaN and without this use of NaN is undefined behavior.
Because of a compiler defect which only occurs with -ffp-mode=full a
workaround is done in the lstm unit test. The workaround prevents
implicit conversion from double to float. Since many of those
conversions would be inexact, the compiler would not consider those
conversions as constant expressions with -ffp-mode=full.

Change-Id: I8de1356b702d8c0f199b08d4a368c4cbe277c670