Comments by Advait :
While attempting to get the seanet NN-only benchmark running we realized that
the NN tflite file has float input and outputs (and corresponding quantize op
    with float input, and dequantize op at the end of the NN).

These are modified to be int32 and int16 (and change the dequantize to quantize)
  as part of the stft_model_merger tool:
  http://google3/third_party/tflite_micro/tensorflow/lite/micro/audio_frontend/python/utils/stft_model_merger_lib.py;l=172-174;rcl=460600505

This is not really the path we should be going down becuase the STFT model
merger should be generic and not based on the quirks of the SEANet model.

For now, this CL is adding quantize with float inputs to the xtensa
implementaiton as well to allow us to run NN-only benchmarks.

Exactly how we want to clean all this up is currently undecided.