enable reduce_sum kernel tests for Xtensa.
PR #1229 ported over reduce sum but did not add a reference fallback to
`kernels/xtensa/reduce.cc`. As a result #1229 put the kernel tests behind
a `#if !defined(XTENSA)` guard.

Instead of excluding targets in the kernel tests, what we would like to
do is add reference code fallbacks to the optimized kernel sources so
that we can guarantee that the optimized kernels have the same
functionality as the reference kernels (even if not all the code paths
are truly optimized).

With PR #1274, we added reference fallbacks to `xtensa/reduce.cc` but
missed the fact that the kernel test cases had been ifdef'd out.

The current PR enables all the kernels test cases for all targets.

BUG=http://b/236681001