Issue on error
New issue creation and updating mechanism for action error trapping.

On error creates a new issue with the tag ci:bot_issue if one doesn't exist. Places information about the script failing and run in the body.

If an issue with the script name in the body and the ci:bot_issue tag exists and is open, will update with a comment about the current failure.

Error trapping implemented for testing in generate_integration_tests.yml

BUG=https://issuetracker.google.com/issues/229828377