Fix TF 1.x deprecation speech training notebook
Google Colabs removed support for the `%tensorflow 1.x` directive in August 2022. This notebook still relies on 1.x functionality, and isn't easy to adapt, so I've suggested a change that switches to the 1.15 CPU version of the framework. This doesn't take advantage of the GPU, so it's a bit slower, but only increases the time to about two hours instead of one in my experiments, so I think it's the least-worst option.