Change the xtensa tests to be able to run from the parent of tflm-micro/ directory.
This is a follow up PR for [1446](https://github.com/tensorflow/tflite-micro/pull/1446). Once [1446](https://github.com/tensorflow/tflite-micro/pull/1446) is merged, this pr is going to run xtensa tests from the  /opt (parent of the tflite-micro directory) and we are doing this to be consistent with our internal CI Build. 

BUG=http://b/244204467