Port REDUCE_ANY OP and generalize Reduce Helper
Port the new op from Lite and add it to the list of all ops
for Micro.

Prepare was implemeneted for REDUCE_MAX port but it is common
for all REDUCE_XXX ops, thus its helper is renamed to
PrepareReduceHelper from PrepareMaxHelper. Similarly, EvalMaxHelper
is renamed to EvalReduceHelper and its implementation body is
generalized to support all REDUCE_XXX ops.