Replace deprecated gen_lut usage by LUTPopulate
Hello,

This PR replaces the usage of the `gen_lut` function, deprecated by the https://github.com/tensorflow/tensorflow/pull/54390 PR, by the `LUTPopulate` one (and use `LUTSize` to calculate the size of the LUT). 

It was kept in the original PR to avoid any difficult conflicts during the merge, removing its usage here would allow us to remove it alltogether in TFLite.

The resulting `exp_lut` and `one_over_one_plus_x_lut` are exactly the same as the ones generated by `gen_lut` and a double-check was done to ensure that. The PR does not change any behaviour.

Thanks,
Thibaut

BUG=https://github.com/tensorflow/tensorflow/pull/54390