build: generalize Py package C-language repository_rule
Generalize the naming and documentation of the repository rule
that creates targets for C-language components shipped inside
Python packages (to be used as dependencies by C-language code)
to prepare for exporting both headers and binary libraries. The
current rule is capable of exporting only headers, and was named
and documented with that limiting assumption.