build(bazel): exclude Python requirements test from wildcard
Exclude the test of the pinned Python requirements file from
wildcard patterns such as that used by CI to execute all tests.
This excluded test ensures that the compiled, pinned requirements
file is up to date with the requirements.in file from which it is
generated.

Exclude this test for now because it is sensitive to the version
of Python under which it is run. If the pinned file is written by
a developer using a different version of Python than that used in
CI, the test will fail.

Perhaps solve in the near future by configuring Bazel to use its
own Python interpreter rather than depending on the interpreter
provided by the system on which it is running.