Optimize Xtensa CONV2D circular buffer copy.
In Xtensa's CONV2D kernel, data is shuffled around and padded so the 2D convolution turns into sequential vector products.
Unfortunately, this process is somewhat slow, and the overhead is especially high for small vector lengths.

This patch introduces the following:

- Faster code path for no padding (since our models use VALID padding, i.e., no padding at all)
- Manual loop if array is small and memcpy if array is large
- Skip memset on padded channels as the corresponding kernels are already zero

This results in a ~10% cycle reduction on SEANet models, and removes the regressions caused by switching UNR layers from TRANSPOSED_CONV to CONV2D.

BUG=249796929