Adding python wrapper for tflite flatbuffer align tool (#1977)
The `tflite_flatbuffer_align` tool is currently a C++ script. The functionality of this script is only useful after you have modified the tflite flatbuffer via the Python API, then you would call this script to realign via the C++ API.

It would make more sense to wrap this tool in python, and use it directly in the python script, instead of having a two step procedure.
We leave a py binary with the same bazel target +args structure to keep support for any prior usecases.

BUG=[274637329](http://b/274637329)