build: remove intermediate pybind_library target
Remove the pybind_library() used only as a dependency to the
pybind_extension(), and instead simply make the
pybind_extension() directly. A pybind_library() may be useful if
used in other palces in addition to a pybind_extension(), but
with no such use in our tree, it just adds complication.

BUG=part of #1484