Use consistent paths in hello world README
The hello world example README.md contained a mix of paths to the bazel
targets, with an implicit assumption that the working directory was
either `tensorflow/lite/micro/examples/hello_world` or the project root.
This CL updates the paths to consistently assume the current working
directory is the project root for clarity.

BUG=doc cleanup