refactor(runtime.py): count resource variables in native code
Move the counting of resource variables from the Python code to
the native code of the soon-to-be-public Python module
tflm_micro.runtime. This avoids what would be a large, complex
dependency in the public module on the third-party packages
tensorflow, tflite-runtime, and flatbuffers (or vendored parts
thereof).

This code is covered by the existing tests.

BUG=part of #1484