Add TF_LITE_ASSERT macro
In upstream TFLite, op_macros.h contains a TF_LITE_ASSERT macro that
would print the failed assertion and then call abort(). This macro was
removed from TFLM's version of op_macros.h because it relied on stdlib
functions (namely, fprintf). As there were no usages of TF_LITE_ASSERT
in shared code, this was not a problem. However, the latest sync
introduces a usage of it within shared code, breaking the build. This
commit adds a TFLM-friendly implementation of TF_LITE_ASSERT using
MicroPrintf for the implementation.